requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
numpy>=1.21.0
opencv-python>=4.5.0
tqdm>=4.62.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
Pillow>=8.3.0
tensorboard>=2.8.0
seaborn>=0.11.0
streamlit>=1.28.0
plotly>=5.0.0
altair>=4.2.0